+++ src/icalendar/tests/test_fixed_issues.py
@@ -28,6 +28,29 @@ class TestIssues(unittest.TestCase):
         self.assertEqual(tz['tzid'].to_ical(), "America/New_York")
 
 
+    def test_issue_55(self):
+        """Issue #55 - Parse error on utc-offset with seconds value
+        https://github.com/collective/icalendar/issues/55
+        """
+        ical_str = """BEGIN:VTIMEZONE
+TZID:America/Los Angeles
+BEGIN:STANDARD
+DTSTART:18831118T120702
+RDATE:18831118T120702
+TZNAME:PST
+TZOFFSETFROM:-075258
+TZOFFSETTO:-0800
+END:STANDARD
+END:VTIMEZONE"""
+
+        tz = icalendar.Timezone.from_ical(ical_str)
+        self.assertEqual(tz.to_ical(),
+            'BEGIN:VTIMEZONE\r\nTZID:America/Los Angeles\r\nBEGIN:STANDARD\r\n'
+            'DTSTART:18831118T120702\r\nRDATE:18831118T120702\r\nTZNAME:PST'
+            '\r\nTZOFFSETFROM:-075258\r\nTZOFFSETTO:-0800\r\nEND:STANDARD\r\n'
+            'END:VTIMEZONE\r\n')
+
+
     def test_issue_58(self):
         """Issue #58 - TZID on UTC DATE-TIMEs
         https://github.com/collective/icalendar/issues/58
@@ -82,8 +105,7 @@ SUMMARY:Esb mellon phone conf
 DTSTART:20070220T170000
 DTSTAMP:20070221T095412Z
 SEQUENCE:0
-END:VEVENT
-"""
+END:VEVENT"""
 
         cal = icalendar.Calendar.from_ical(ical_str)
         recur = cal.decoded("RRULE")